-Sorted internal classes. -Added more exception messages for parsing errors.
