+++ src/MissingValues.Tests/Core/UInt256Test.GenericMath.cs
@@ -962,7 +962,7 @@ namespace MissingValues.Tests.Core
 				.Should().Be(MaxValue)
 				.And.BeRankedEquallyTo(MaxValue);
 
-			Assert.Throws<FormatException>(() => NumberBaseHelper<UInt>.Parse("115792089237316195423570985008687907853269984665640564039457584007913129639936", System.Globalization.NumberStyles.Integer, CultureInfo.CurrentCulture));
+			Assert.Throws<OverflowException>(() => NumberBaseHelper<UInt>.Parse("115792089237316195423570985008687907853269984665640564039457584007913129639936", System.Globalization.NumberStyles.Integer, CultureInfo.CurrentCulture));
 		}
 
 		[Fact]
@@ -992,7 +992,7 @@ namespace MissingValues.Tests.Core
 				.Should().Be(MaxValue)
 				.And.BeRankedEquallyTo(MaxValue);
 
-			Assert.Throws<FormatException>(() => NumberBaseHelper<UInt>.Parse("115792089237316195423570985008687907853269984665640564039457584007913129639936"u8, System.Globalization.NumberStyles.Integer, CultureInfo.CurrentCulture));
+			Assert.Throws<OverflowException>(() => NumberBaseHelper<UInt>.Parse("115792089237316195423570985008687907853269984665640564039457584007913129639936"u8, System.Globalization.NumberStyles.Integer, CultureInfo.CurrentCulture));
 		}
 
 		[Fact]
