+++ src/MissingValues/Int256.Implementations.cs
@@ -299,32 +299,64 @@ namespace MissingValues
 
 		public static Int256 Parse(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider? provider)
 		{
-			return NumberParser.ParseToSigned<Int256, UInt256, Utf16Char>(Utf16Char.CastFromCharSpan(s), style, provider);
+			var status = NumberParser.TryParseToSigned<Int256, UInt256, Utf16Char>(Utf16Char.CastFromCharSpan(s), style, provider, out Int256 output);
+			if (!status)
+			{
+				status.Throw<Int256>(s.ToString());
+			}
+			return output;
 		}
 
 		public static Int256 Parse(string s, NumberStyles style, IFormatProvider? provider)
 		{
-			return NumberParser.ParseToSigned<Int256, UInt256, Utf16Char>(Utf16Char.CastFromCharSpan(s), style, provider);
+			var status = NumberParser.TryParseToSigned<Int256, UInt256, Utf16Char>(Utf16Char.CastFromCharSpan(s), style, provider, out Int256 output);
+			if (!status)
+			{
+				status.Throw<Int256>(s.ToString());
+			}
+			return output;
 		}
 
 		public static Int256 Parse(ReadOnlySpan<char> s, IFormatProvider? provider)
 		{
-			return NumberParser.ParseToSigned<Int256, UInt256, Utf16Char>(Utf16Char.CastFromCharSpan(s), NumberStyles.Integer, provider);
+			var status = NumberParser.TryParseToSigned<Int256, UInt256, Utf16Char>(Utf16Char.CastFromCharSpan(s), NumberStyles.Integer, provider, out Int256 output);
+			if (!status)
+			{
+				status.Throw<Int256>(s.ToString());
+			}
+			return output;
 		}
 
 		public static Int256 Parse(string s, IFormatProvider? provider)
 		{
-			return NumberParser.ParseToSigned<Int256, UInt256, Utf16Char>(Utf16Char.CastFromCharSpan(s), NumberStyles.Integer, provider);
+			var status = NumberParser.TryParseToSigned<Int256, UInt256, Utf16Char>(Utf16Char.CastFromCharSpan(s), NumberStyles.Integer, provider, out Int256 output);
+			if (!status)
+			{
+				status.Throw<Int256>(s.ToString());
+			}
+			return output;
 		}
 
 #if NET8_0_OR_GREATER
 		public static Int256 Parse(ReadOnlySpan<byte> utf8Text, NumberStyles style, IFormatProvider? provider)
 		{
-			return NumberParser.ParseToSigned<Int256, UInt256, Utf8Char>(Utf8Char.CastFromByteSpan(utf8Text), style, provider);
+			var status = NumberParser.TryParseToSigned<Int256, UInt256, Utf8Char>(Utf8Char.CastFromByteSpan(utf8Text), style, provider, out Int256 output);
+			if (!status)
+			{
+				string input = new string(Encoding.UTF8.GetChars(utf8Text.ToArray()));
+				status.Throw<Int256>(input);
+			}
+			return output;
 		}
 		public static Int256 Parse(ReadOnlySpan<byte> utf8Text, IFormatProvider? provider)
 		{
-			return NumberParser.ParseToSigned<Int256, UInt256, Utf8Char>(Utf8Char.CastFromByteSpan(utf8Text), NumberStyles.Integer, provider);
+			var status = NumberParser.TryParseToSigned<Int256, UInt256, Utf8Char>(Utf8Char.CastFromByteSpan(utf8Text), NumberStyles.Integer, provider, out Int256 output);
+			if (!status)
+			{
+				string input = new string(Encoding.UTF8.GetChars(utf8Text.ToArray()));
+				status.Throw<Int256>(input);
+			}
+			return output;
 		}
 #endif
 
@@ -902,18 +934,18 @@ namespace MissingValues
 
 		public string ToString(string? format, IFormatProvider? formatProvider)
 		{
-			return NumberFormatter.FormatSignedNumber<Int256, UInt256>(in this, format, NumberStyles.Integer, formatProvider);
+			return NumberFormatter.FormatSignedInteger<Int256, UInt256>(in this, format, NumberStyles.Integer, formatProvider);
 		}
 
 		public bool TryFormat(Span<char> destination, out int charsWritten, ReadOnlySpan<char> format, IFormatProvider? provider)
 		{
-			return NumberFormatter.TryFormatSignedNumber<Int256, UInt256, Utf16Char>(in this, Utf16Char.CastFromCharSpan(destination), out charsWritten, format, provider);
+			return NumberFormatter.TryFormatSignedInteger<Int256, UInt256, Utf16Char>(in this, Utf16Char.CastFromCharSpan(destination), out charsWritten, format, provider);
 		}
 
 #if NET8_0_OR_GREATER
 		public bool TryFormat(Span<byte> utf8Destination, out int bytesWritten, ReadOnlySpan<char> format, IFormatProvider? provider)
 		{
-			return NumberFormatter.TryFormatSignedNumber<Int256, UInt256, Utf8Char>(in this, Utf8Char.CastFromByteSpan(utf8Destination), out bytesWritten, format, provider);
+			return NumberFormatter.TryFormatSignedInteger<Int256, UInt256, Utf8Char>(in this, Utf8Char.CastFromByteSpan(utf8Destination), out bytesWritten, format, provider);
 		} 
 #endif
 
@@ -1015,7 +1047,7 @@ namespace MissingValues
 			return (int)this;
 		}
 
-		static Int256 IFormattableInteger<Int256>.GetDecimalValue(char value)
+		static Int256 IFormattableNumber<Int256>.GetDecimalValue(char value)
 		{
 			if (!char.IsDigit(value))
 			{
