+++ src/MissingValues/Internals/BitHelper.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using MissingValues.Internals;
+using System;
 using System.Buffers.Binary;
 using System.Collections.Generic;
 using System.Linq;
