+++ src/MissingValues/Internals/IFormattableNumber.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System.Numerics;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MissingValues.Internals
+{
+	internal interface IFormattableNumber<TSelf> : INumber<TSelf>
+		where TSelf : IFormattableNumber<TSelf>?
+	{
+		/// <summary>
+		/// Converts the specified numeric Unicode character to a <typeparamref name="TSelf"/>.
+		/// </summary>
+		/// <param name="value">A numeric Unicode character.</param>
+		/// <returns>The numeric value of <paramref name="value"/> if it represents a number; otherwise, 0</returns>
+		virtual static TSelf GetDecimalValue(char value)
+		{
+			return TSelf.CreateChecked(CharUnicodeInfo.GetDecimalDigitValue(value));
+		}
+
+		abstract static bool IsBinaryInteger();
+	}
+}
