+++ src/MissingValues/UInt256.Implementations.cs
@@ -16,7 +16,7 @@ using System.Threading.Tasks;
 
 namespace MissingValues
 {
-    public readonly partial struct UInt256 : 
+	public readonly partial struct UInt256 : 
 		IBinaryInteger<UInt256>,
 		IMinMaxValue<UInt256>,
 		IUnsignedNumber<UInt256>,
@@ -239,32 +239,62 @@ namespace MissingValues
 
 		public static UInt256 Parse(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider? provider)
 		{
-			return NumberParser.ParseToUnsigned<UInt256, Utf16Char>(Utf16Char.CastFromCharSpan(s), style, provider);
+			var status = NumberParser.TryParseToUnsigned(Utf16Char.CastFromCharSpan(s), style, provider, out UInt256 output);
+			if (!status)
+			{
+				status.Throw<UInt256>(s.ToString());
+			}
+			return output;
 		}
 
 		public static UInt256 Parse(string s, NumberStyles style, IFormatProvider? provider)
 		{
-			return NumberParser.ParseToUnsigned<UInt256, Utf16Char>(Utf16Char.CastFromCharSpan(s), style, provider);
+			var status = NumberParser.TryParseToUnsigned(Utf16Char.CastFromCharSpan(s), style, provider, out UInt256 output);
+			if (!status)
+			{
+				status.Throw<UInt256>(s.ToString());
+			}
+			return output;
 		}
 
 		public static UInt256 Parse(ReadOnlySpan<char> s, IFormatProvider? provider)
 		{
-			return NumberParser.ParseToUnsigned<UInt256, Utf16Char>(Utf16Char.CastFromCharSpan(s), NumberStyles.Integer, provider);
+			var status = NumberParser.TryParseToUnsigned(Utf16Char.CastFromCharSpan(s), NumberStyles.Integer, provider, out UInt256 output);
+			if (!status)
+			{
+				status.Throw<UInt256>(s.ToString());
+			}
+			return output;
 		}
 
 		public static UInt256 Parse(string s, IFormatProvider? provider)
 		{
-			return NumberParser.ParseToUnsigned<UInt256, Utf16Char>(Utf16Char.CastFromCharSpan(s), NumberStyles.Integer, provider);
+			var status = NumberParser.TryParseToUnsigned(Utf16Char.CastFromCharSpan(s), NumberStyles.Integer, provider, out UInt256 output);
+			if (!status)
+			{
+				status.Throw<UInt256>(s.ToString());
+			}
+			return output;
 		}
 
 #if NET8_0_OR_GREATER
 		public static UInt256 Parse(ReadOnlySpan<byte> utf8Text, NumberStyles style, IFormatProvider? provider)
 		{
-			return NumberParser.ParseToUnsigned<UInt256, Utf8Char>(Utf8Char.CastFromByteSpan(utf8Text), style, provider);
+			var status = NumberParser.TryParseToUnsigned(Utf8Char.CastFromByteSpan(utf8Text), style, provider, out UInt256 output);
+			if (!status)
+			{
+				status.Throw<UInt256>(utf8Text);
+			}
+			return output;
 		}
 		public static UInt256 Parse(ReadOnlySpan<byte> utf8Text, IFormatProvider? provider)
 		{
-			return NumberParser.ParseToUnsigned<UInt256, Utf8Char>(Utf8Char.CastFromByteSpan(utf8Text), NumberStyles.Integer, provider);
+			var status = NumberParser.TryParseToUnsigned(Utf8Char.CastFromByteSpan(utf8Text), NumberStyles.Integer, provider, out UInt256 output);
+			if (!status)
+			{
+				status.Throw<UInt256>(utf8Text);
+			}
+			return output;
 		}
 #endif
 
@@ -701,13 +731,13 @@ namespace MissingValues
 
 		public bool TryFormat(Span<char> destination, out int charsWritten, [StringSyntax(StringSyntaxAttribute.NumericFormat)] ReadOnlySpan<char> format, IFormatProvider? provider)
 		{
-			return NumberFormatter.TryFormatUnsignedNumber(in this, Utf16Char.CastFromCharSpan(destination), out charsWritten, format, provider);
+			return NumberFormatter.TryFormatUnsignedInteger(in this, Utf16Char.CastFromCharSpan(destination), out charsWritten, format, provider);
 		}
 
 #if NET8_0_OR_GREATER
 		public bool TryFormat(Span<byte> utf8Destination, out int bytesWritten, [StringSyntax(StringSyntaxAttribute.NumericFormat)] ReadOnlySpan<char> format, IFormatProvider? provider)
 		{
-			return NumberFormatter.TryFormatUnsignedNumber(in this, Utf8Char.CastFromByteSpan(utf8Destination), out bytesWritten, format, provider);
+			return NumberFormatter.TryFormatUnsignedInteger(in this, Utf8Char.CastFromByteSpan(utf8Destination), out bytesWritten, format, provider);
 		}
 #endif
 
@@ -804,7 +834,7 @@ namespace MissingValues
 			return (int)this;
 		}
 
-		static UInt256 IFormattableInteger<UInt256>.GetDecimalValue(char value)
+		static UInt256 IFormattableNumber<UInt256>.GetDecimalValue(char value)
 		{
 			if (!char.IsDigit(value))
 			{
