+++ src/MissingValues/UInt256.cs
@@ -141,7 +141,7 @@ namespace MissingValues
 
 		public override string ToString()
 		{
-			return NumberFormatter.UnsignedNumberToString(in this, new UInt256(UInt128.Zero, 10));
+			return NumberFormatter.UnsignedIntegerToString(in this, new UInt256(UInt128.Zero, 10));
 		}
 
 		public override bool Equals(object? obj)
